Some changes
- Broke the project down to multiple files
- Added shutdown functionality
- Added more safety
- Now using channels instead of direct logging
- Some tests have been done
- Updated readme
